feat : #22 - getDragAfterElement, separateDiv, appendDiv 파일 분리 완료
