+++ src/utils/index.js
@@ -2,3 +2,5 @@ export { areEqual } from "./areEqual";
 export { debounce } from "./debounce";
 export { executeWithTryCatch } from "./excuteWithTryCatch";
 export { setCaretOffset, saveCursor, lastCursor } from "./Cursor";
+export { appendDiv, separateDiv } from "./controlEditorDiv";
+export { getDragAfterElement } from "./controlEditorDrag";
